chore: Update Starter project to fit latest version which was on a different folder
