+++ Starter-project/gradle.properties
@@ -44,10 +44,10 @@ android.privacySandboxSdk.apiPackager=androidx.privacysandbox.tools:tools:1.0.0-
   androidx.privacysandbox.tools:tools-apipackager:1.0.0-alpha13
 android.privacySandboxSdk.apiGenerator.generatedRuntimeDependencies=\
   androidx.privacysandbox.tools:tools-apigenerator:1.0.0-alpha13,\
-  org.jetbrains.kotlin:kotlin-stdlib:2.1.10,\
-  org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3,\
+  org.jetbrains.kotlin:kotlin-stdlib:2.2.0,\
+  org.jetbrains.kotlinx:kotlinx-coroutines-android:1.10.2,\
   androidx.privacysandbox.activity:activity-core:1.0.0-alpha02,\
   androidx.privacysandbox.activity:activity-client:1.0.0-alpha02,\
   androidx.privacysandbox.activity:activity-provider:1.0.0-alpha02,\
-  androidx.privacysandbox.ui:ui-core:1.0.0-alpha15,\
-  androidx.privacysandbox.ui:ui-client:1.0.0-alpha15
+  androidx.privacysandbox.ui:ui-core:1.0.0-alpha16,\
+  androidx.privacysandbox.ui:ui-client:1.0.0-alpha16
