+++ Starter-project/gradle/libs.versions.toml
@@ -17,20 +17,20 @@
 # or SDK Runtime Jetpack library versions (Shim tools, Backcompat sdkrt, ui, activity),
 # make sure to also update the dependencies specified in `gradle.properties`.
 activityCompose = "1.10.1"
-composeBom = "2025.06.00"
-kotlin = "2.1.20"
-ksp = "2.1.20-1.0.31"
+composeBom = "2025.06.01"
+kotlin = "2.2.0"
+ksp = "2.2.0-2.0.2"
 agp = "8.9.2"
 privacySandboxBuildPlugin = "1.0.0-alpha02"
 sdkRuntimeBackcompat = "1.0.0-alpha17"
 sdkRuntimeShim = "1.0.0-alpha13"
-sdkRuntimeUi = "1.0.0-alpha15"
+sdkRuntimeUi = "1.0.0-alpha16"
 sdkRuntimeActivity = "1.0.0-alpha02"
-appcompat = "1.7.0"
+appcompat = "1.7.1"
 material = "1.12.0"
 activityKtx = "1.10.1"
-lifecycleCommon = "2.8.7"
-coroutines = "1.7.3"
+lifecycleCommon = "2.9.1"
+coroutines = "1.10.2"
 annotation = "1.9.1"
 kotlinStdlibJdk8 = "2.1.10"
 androidxTestRunner = "1.6.2"
